requirements.txt
numpy
matplotlib
Pillow
torch
torchvision
